Forbedringer på router og struktur på filer/folders
